🐛fix : 거리는 미터 단위 int, 시간은 초 단위 int로 변경

+++ backend/src/main/java/com/luckycookie/crewin/dto/CourseRequest.java
@@ -27,7 +27,7 @@ public class CourseRequest {
         private String info;
         private String name;
         private String area;
-        private Double length;
+        private Integer length;
         private String thumbnailImage ;
     }
